+++ src/policy/policy.cpp
@@ -52,6 +52,71 @@ bool IsStandard(const Config &config, const CScript &scriptPubKey, int nScriptPu
     return whichType != TX_NONSTANDARD;
 }
 
+// Check if a transaction is a consolidation transaction.
+// A consolidation transaction is a transaction which reduces the size of the UTXO database to
+// an extent that is rewarding enough for the miner to mine the transaction for free.
+bool IsConsolidationTxn(const Config &config, const CTransaction &tx, const CCoinsViewCache &inputs, int tipHeight)
+{
+    const uint64_t factor = config.GetMinConsolidationFactor();
+    const uint64_t minMaturity = config.GetMinConsolidationInputMaturity();
+    const uint64_t maxSize = config.GetMaxConsolidationInputScriptSize();
+    const bool stdInputOnly = !config.GetAcceptNonStdConsolidationInput();
+
+
+    // Allow disabling free consolidation txns via configuring
+    // the consolidation factor to zero
+    if (factor == 0)
+        return false;
+
+    if (tx.IsCoinBase())
+        return false;
+
+    // The consolidation transaction needs to reduce the count of UTXOS
+    if (tx.vin.size() < factor * tx.vout.size())
+        return false;
+
+    // Check all UTXOs are confirmed and prevent spam via big
+    // scriptSig sizes in the consolidation transaction inputs.
+    uint64_t sumScriptPubKeySizeOfTxInputs = 0;
+    for (CTxIn const & u: tx.vin) {
+
+        // accept only with many confirmations
+        const Coin &coin = inputs.AccessCoin(u.prevout);
+        const auto coinHeight = coin.GetHeight();
+
+        if (coinHeight == MEMPOOL_HEIGHT)
+            return false;
+
+        if (coinHeight && (tipHeight - coinHeight < minMaturity)) // older versions did not store height
+            return false;
+
+        // spam detection
+        if (u.scriptSig.size() > maxSize)
+            return false;
+
+        // if not acceptnonstdconsolidationinput then check if inputs are standard
+        // and fail otherwise
+        txnouttype dummyType;
+        if (stdInputOnly  && !IsStandard(config, coin.GetTxOut().scriptPubKey, coinHeight, dummyType))
+            return false;
+
+        // sum up some script sizes
+        sumScriptPubKeySizeOfTxInputs += coin.GetTxOut().scriptPubKey.size();
+    }
+
+    // check ratio between sum of tx-scriptPubKeys to sum of parent-scriptPubKeys
+    uint64_t sumScriptPubKeySizeOfTxOutputs = 0;
+    for (CTxOut const & o: tx.vout) {
+        sumScriptPubKeySizeOfTxOutputs += o.scriptPubKey.size();
+    }
+
+    // prevent consolidation transactions that are not advantageous enough for miners
+    if(sumScriptPubKeySizeOfTxInputs < factor * sumScriptPubKeySizeOfTxOutputs)
+        return false;
+
+    return true;
+}
+
 bool IsStandardTx(const Config &config, const CTransaction &tx, int nHeight, std::string &reason) {
     if (tx.nVersion > CTransaction::MAX_STANDARD_VERSION || tx.nVersion < 1) {
         reason = "version";