+++ src/rpc/net.cpp
@@ -602,6 +602,11 @@ static UniValue getnetworkinfo(const Config &config,
         obj.push_back(Pair("connections", (int)g_connman->GetNodeCount(CConnman::CONNECTIONS_ALL)));
         obj.push_back(Pair("addresscount", static_cast<uint64_t>(g_connman->GetAddressCount())));
     }
+    obj.push_back(Pair("minconsolidationfactor",  config.GetMinConsolidationFactor()));
+    obj.push_back(Pair("maxconsolidationinputscriptsize",  config.GetMaxConsolidationInputScriptSize()));
+    obj.push_back(Pair("minconsolidationinputmaturity",  config.GetMinConsolidationInputMaturity()));
+    obj.push_back(Pair("acceptnonstdconsolidationinput",  config.GetAcceptNonStdConsolidationInput()));
+
     obj.push_back(Pair("networks", GetNetworksInfo()));
     obj.push_back(Pair("relayfee",
                        ValueFromAmount(config.GetMinFeePerKB().GetFeePerK())));